+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Moved module to src folder
+- Move all options onto an options sub-object
+
+### Added
+- UMD variant to umd folder
+- Add "waitMillis" and "waitFrames" options
+
 ## [0.0.2] - 2018-08-26
 ### Fixed
 - Properly remove events from window on dispose.